+++ quantforge/__init__.py
@@ -781,6 +781,7 @@ from .tucker import tucker_hosvd, tucker_reconstruct
 from .spectral_clustering import spectral_clustering, rbf_affinity
 from .mds import classical_mds, embedded_distances
 from .isomap import isomap, knn_graph
+from .laplacian_eigenmaps import laplacian_eigenmaps, heat_knn_affinity
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1485,7 +1486,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "5.76.0"
+__version__ = "5.77.0"
 
 __all__ = [
     "OptionType",
@@ -2541,6 +2542,8 @@ __all__ = [
     "embedded_distances",
     "isomap",
     "knn_graph",
+    "laplacian_eigenmaps",
+    "heat_knn_affinity",
     "Dual",
     "dual_derivative",
     "dual_gradient",